+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2020-07-09
+
+### Changed
+
+- Progress bars now work in Jupyter
+
+## Added
+
+- Added refresh_per_second to progress.track
+- Added styles to BarColumn and progress.track
+
 ## [3.0.5] - 2020-07-07
 
 ### Fixed